move side nav component code to library package
