+++ src/renderer/library/utils/Wrappers.ts
@@ -1,0 +1,24 @@
+// Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+import styled from 'styled-components';
+
+export const ScrollWrapper = styled.div`
+  scrollbar-color: inherit transparent;
+
+  overflow-y: auto;
+  overflow-x: hidden;
+  position: relative;
+
+  -ms-overflow-style: none;
+
+  &::-webkit-scrollbar {
+    width: 5px;
+  }
+  &::-webkit-scrollbar-track {
+    background-color: #101010;
+  }
+  &::-webkit-scrollbar-thumb {
+    background-color: #212121;
+  }
+`;
